// types.ts
export interface Customer {
  id: string;
  name: string;
  email: string;
  phone: string;
  address: string;
  tier: 'Bronze' | 'Silver' | 'Gold' | 'Platinum';
  rewardPoints: number;
  totalSpent: number;
  wishlistIds: string[];
  ordersCount: number;
  joinedDate: string;
}

export type UserRole = 'Admin' | 'Manager' | 'Cashier' | 'Kitchen' | 'Customer';
// ...
export interface SystemUser {
  id: string;
  username: string;
  fullName: string;
  role: 'Admin' | 'Manager' | 'Cashier' | 'Kitchen';
  passwordHash: string;
  allowedModules: string[]; // List of ERP module IDs allowed for this user
  avatarUrl?: string;
  isActive: boolean;
}